Made the chart a react component so it looks much neater
